update rust control
