+++ meta-cross/recipes-rust/controller/files/car_controller/controller/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "controller"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+vehicle_control = { path = "../lib/vehicle_control" }
